requirements.txt
flask
flask-cors
flask-migrate
flask-restful
flask-sqlalchemy
python-dotenv
